backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "aiofiles==25.1.0",
    "aiohttp==3.13.5",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.13.0",
    "async-timeout==5.0.1",
    "attrs==26.1.0",
    "certifi==2026.2.25",
    "charset-normalizer==3.4.7",
    "click==8.3.2",
    "colorama==0.4.6",
    "exceptiongroup==1.3.1",
    "faiss-cpu==1.13.2",
    "fastapi==0.135.3",
    "filelock==3.25.2",
    "frozenlist==1.8.0",
    "fsspec==2026.3.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface-hub==1.9.0",
    "idna==3.11",
    "jinja2==3.1.6",
    "joblib==1.5.3",
    "langchain>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.1.0",
    "markupsafe==3.0.3",
    "mftool>=2.0",
    "mpmath==1.3.0",
    "multidict==6.7.1",
    "networkx==3.6.1",
    "nltk==3.9.4",
    "numpy==1.26.4",
    "packaging==26.0",
    "pdfplumber==0.11.4",
    "pillow==12.2.0",
    "propcache==0.4.1",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.13.1",
    "pymupdf==1.25.1",
    "pytesseract==0.3.13",
    "python-dotenv==1.2.2",
    "python-multipart==0.0.22",
    "pyyaml==6.0.2",
    "qdrant-client==1.17.1",
    "regex==2026.4.4",
    "requests==2.33.1",
    "safetensors==0.7.0",
    "scikit-learn==1.8.0",
    "scipy==1.17.1",
    "sentence-transformers==5.3.0",
    "sentencepiece==0.2.1",
    "sniffio==1.3.1",
    "starlette==1.0.0",
    "sympy==1.14.0",
    "threadpoolctl==3.6.0",
    "tokenizers==0.22.2",
    "torch==2.11.0",
    "torchvision==0.26.0",
    "tqdm==4.67.3",
    "transformers==5.5.0",
    "typing-extensions==4.15.0",
    "urllib3==2.6.3",
    "uvicorn==0.43.0",
    "yarl==1.23.0",
]
